+++ src/main/java/com/bytehonor/sdk/intent/human/HumanIntentRecognizer.java
@@ -22,7 +22,6 @@ import com.bytehonor.sdk.intent.human.resolver.AskAbilityIntentResolver;
 import com.bytehonor.sdk.intent.human.resolver.AskAgeIntentResolver;
 import com.bytehonor.sdk.intent.human.resolver.AskMusicIntentResolver;
 import com.bytehonor.sdk.intent.human.resolver.AskNameIntentResolver;
-import com.bytehonor.sdk.intent.human.resolver.FinishIntentResolver;
 import com.bytehonor.sdk.intent.human.resolver.IntentResolver;
 import com.bytehonor.sdk.intent.human.resolver.UnsupportIntentResolver;
 import com.bytehonor.sdk.intent.human.worker.IntentWorker;
@@ -33,8 +32,6 @@ import com.bytehonor.sdk.lang.spring.util.JacksonUtils;
 public final class HumanIntentRecognizer {
 
     private static final Logger LOG = LoggerFactory.getLogger(HumanIntentRecognizer.class);
-
-    private static final String END_NAME = FinishIntentResolver.class.getSimpleName();
 
     private final IntentContext context;
 
@@ -63,7 +60,6 @@ public final class HumanIntentRecognizer {
         recognizer.add(new AskAbilityIntentResolver());
         recognizer.add(new AskNameIntentResolver());
         recognizer.add(new AskAgeIntentResolver());
-        recognizer.add(new FinishIntentResolver());
         return recognizer;
     }
 
@@ -87,7 +83,6 @@ public final class HumanIntentRecognizer {
         worker.put(session.getUuid(), session);
 
         result.setSession(session);
-        result.setFinished(END_NAME.equals(result.getResolver()));
 
         IntentListenerThread.add(result);
     }
