+++ src/main/java/com/bytehonor/sdk/intent/human/matcher/IntentMatcher.java
@@ -10,10 +10,8 @@ import org.springframework.util.CollectionUtils;
 
 import com.bytehonor.sdk.lang.spring.constant.CharConstants;
 import com.bytehonor.sdk.lang.spring.match.WordMatcher;
-import com.bytehonor.sdk.lang.spring.nlp.TextNlpUtils;
 import com.bytehonor.sdk.lang.spring.regex.PatternUtils;
 import com.bytehonor.sdk.lang.spring.string.SpringString;
-import com.bytehonor.sdk.lang.spring.string.StringRemoveUtils;
 import com.bytehonor.sdk.lang.spring.string.StringSliceUtils;
 import com.bytehonor.sdk.lang.spring.string.StringSplitUtils;
 
@@ -87,7 +85,7 @@ public class IntentMatcher {
      */
     public static Set<String> words(String text) {
         // 只保留 英文,数字,汉字,空格
-        text = prepare(text);
+        text = IntentQuery.prepare(text);
         if (SpringString.isEmpty(text)) {
             return new HashSet<String>();
         }
@@ -128,52 +126,6 @@ public class IntentMatcher {
             }
         }
         return result;
-    }
-
-    /**
-     * 只保留 英文,汉字,数字,空格 4种. 非中文之间插入空格
-     * 
-     * @param text
-     * @return
-     */
-    public static String prepare(String text) {
-        text = TextNlpUtils.removeHttp(text);
-        text = StringRemoveUtils.replaceNonNormalWithBlank(text);
-        if (SpringString.isEmpty(text)) {
-            return "";
-        }
-        int length = text.length();
-        if (length < WORD_MIN_LENGTH) {
-            return text;
-        }
-        char[] source = text.toCharArray();
-        char[] target = new char[length * 2];
-        int at = 0;
-        for (int i = 0; i < length; i++) {
-            if (PatternUtils.isNormalChar(source[i])) {
-                if (i > 1 && blank(source[i - 1], source[i])) {
-                    // 非中文之间补个空格
-                    target[at++] = CharConstants.BLANK;
-                }
-                target[at++] = source[i];
-                continue;
-            }
-            target[at++] = CharConstants.BLANK;
-        }
-        return new String(target, 0, at);
-    }
-
-    private static boolean blank(char last, char now) {
-        if (last == CharConstants.BLANK || now == CharConstants.BLANK) {
-            return false;
-        }
-        if (PatternUtils.isChineseChar(last) && PatternUtils.isChineseChar(now) == false) {
-            return true;
-        }
-        if (PatternUtils.isChineseChar(last) == false && PatternUtils.isChineseChar(now)) {
-            return true;
-        }
-        return false;
     }
 
     private static boolean doMatch(Set<String> words, List<WordMatcher> matchers) {
